backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
celery==5.3.4
boto3==1.34.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-docx==1.1.0
PyPDF2==3.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
requests==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0